chore: deport versioning to FichePicture
