+++ src/components/FichePicture.tsx
@@ -1,7 +1,39 @@
 "use client";
+import { useEffect, useState } from "react";
+
 import { fr } from "@codegouvfr/react-dsfr";
+import { format } from "date-fns/format";
 import { PlaceholderValue } from "next/dist/shared/lib/get-img-props";
 import Image from "next/image";
+
+function isRelativeUrl(url) {
+    const absolutePattern = /^(?:[a-zA-Z][a-zA-Z\d+\-.]*:|\/\/)/;
+    return !absolutePattern.test(url);
+}
+
+const computeVersion = () => {
+    const version = format(new Date(), "yyyyMMdd'T'HHmmss");
+    return version;
+};
+
+function isBase64(src: string): boolean {
+    return /^data:image\/[a-z]+;base64,/.test(src);
+}
+
+function addVersionParam(url: string): string {
+    if (typeof window === "undefined") {
+        return url;
+    }
+    // when window is define window.location.origin exists
+    const isRelative = isRelativeUrl(url);
+    const tempUrl = isRelative
+        ? new URL(url, window.location.origin)
+        : new URL(url);
+    const params = tempUrl.searchParams;
+    const version = computeVersion();
+    params.set("v", version);
+    return isRelative ? tempUrl.pathname + tempUrl.search : tempUrl.toString();
+}
 
 export const FichePicture = ({
     src,
@@ -20,6 +52,15 @@ export const FichePicture = ({
     placeholder?: PlaceholderValue;
     initials?: string;
 }) => {
+    const [versionedUrl, setVersionedUrl] = useState<string | null>(null);
+    useEffect(() => {
+        // apply addVersion only in front : prevent hydratation mismatch
+        if (src && !isBase64(src)) {
+            setVersionedUrl(addVersionParam(src));
+        } else {
+            setVersionedUrl(null);
+        }
+    }, [src]);
     const style: React.CSSProperties =
         shape === "round"
             ? {
@@ -48,7 +89,7 @@ export const FichePicture = ({
             {((src || placeholder) && (
                 <Image
                     // @ts-ignore TODO: make TS happy
-                    src={src || placeholder}
+                    src={versionedUrl || src || placeholder}
                     priority={false}
                     sizes="(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw"
                     placeholder={placeholder}
